rename v1 package paths to foo, to avoid semantic import version implications
