+++ fixtures/dup_packages/a/a.go
@@ -1,8 +1,8 @@
 package a
 
-import "github.com/maxbrunsfeld/counterfeiter/fixtures/dup_packages/a/v1"
+import "github.com/maxbrunsfeld/counterfeiter/fixtures/dup_packages/a/foo"
 
 //go:generate counterfeiter . A
 type A interface {
-	V1() v1.I
+	V1() foo.I
 }
