+++ fixtures/dup_packages/dupA.go
@@ -1,8 +1,8 @@
 package dup_packages
 
-import "github.com/maxbrunsfeld/counterfeiter/fixtures/dup_packages/a/v1"
+import "github.com/maxbrunsfeld/counterfeiter/fixtures/dup_packages/a/foo"
 
 //go:generate counterfeiter . DupA
 type DupA interface {
-	A() v1.S
+	A() foo.S
 }
