+++ fixtures/dup_packages/dupB.go
@@ -1,8 +1,8 @@
 package dup_packages
 
-import "github.com/maxbrunsfeld/counterfeiter/fixtures/dup_packages/b/v1"
+import "github.com/maxbrunsfeld/counterfeiter/fixtures/dup_packages/b/foo"
 
 //go:generate counterfeiter . DupB
 type DupB interface {
-	B() v1.S
+	B() foo.S
 }
